Changed index.js, added files to "img" folder, created "classes.js".
Re did the whole index.js to fit the new game that we have done, also added a few pictures to the "img" folder to fit the new code. Created a new file called "classes.js". In the classes folder we will add different classes that are going to be used in the project. Also created a folder where we can stash all our data for things such as levels.