Advance from level to level
Added some lines of code that makes it possible for our character to walk into a door, fade the screen black then fade back again to reveal the new map. Also edited the first map so that there is in fact a door to go through at the end. Added a new js file called "doors" that stores the array of the doors collision block to make the new lines of code work.