Map lvl3 and lvl4 changes
Made changes to the maps lvl 3 and lvl4. Which includes collision blocks, the images were not working properly when going to the sides of the map with the character (pixels and collors moving crazy) but it is fixed now. I also made design changes to the map on the "outside". And changed where the player is spawning when entering these maps.